+++ weblate/api/tests.py
@@ -4,7 +4,7 @@
 
 import os
 from copy import copy
-from datetime import datetime, timedelta, timezone
+from datetime import UTC, datetime, timedelta
 from io import BytesIO
 
 from django.core.files import File
@@ -237,7 +237,9 @@ class UserAPITest(APIBaseTest):
             code=200,
             request={"group_id": group.id},
         )
-        self.assertIn("http://example.com/api/groups/2/", response.data["groups"])
+        self.assertIn(
+            f"http://example.com/api/groups/{group.id}/", response.data["groups"]
+        )
         response = self.do_request(
             "api:user-groups",
             kwargs={"username": username},
@@ -246,7 +248,9 @@ class UserAPITest(APIBaseTest):
             code=200,
             request={"group_id": group.id},
         )
-        self.assertNotIn("http://example.com/api/groups/2/", response.data["groups"])
+        self.assertNotIn(
+            "http://example.com/api/groups/{group.id}/", response.data["groups"]
+        )
 
     def test_list_notifications(self) -> None:
         response = self.do_request(
@@ -1985,8 +1989,8 @@ class ProjectAPITest(APIBaseTest):
             "api:component-credits", self.component_kwargs, method="get", code=400
         )
 
-        start = datetime.now(tz=timezone.utc) - timedelta(days=1)
-        end = datetime.now(tz=timezone.utc) + timedelta(days=1)
+        start = datetime.now(tz=UTC) - timedelta(days=1)
+        end = datetime.now(tz=UTC) + timedelta(days=1)
 
         response = self.do_request(
             "api:component-credits",
@@ -2311,8 +2315,8 @@ class ComponentAPITest(APIBaseTest):
             "api:component-credits", self.component_kwargs, method="get", code=400
         )
 
-        start = datetime.now(tz=timezone.utc) - timedelta(days=1)
-        end = datetime.now(tz=timezone.utc) + timedelta(days=1)
+        start = datetime.now(tz=UTC) - timedelta(days=1)
+        end = datetime.now(tz=UTC) + timedelta(days=1)
 
         response = self.do_request(
             "api:component-credits",