+++ weblate/trans/autofixes/chars.py
@@ -2,18 +2,25 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
+from __future__ import annotations
+
 import re
+from typing import TYPE_CHECKING
 
 from django.utils.translation import gettext_lazy
 
 from weblate.checks.chars import (
-    FRENCH_PUNCTUATION_FIXUP_RE,
+    FRENCH_PUNCTUATION_FIXUP_RE_NBSP,
+    FRENCH_PUNCTUATION_FIXUP_RE_NNBSP,
     EndEllipsisCheck,
     PunctuationSpacingCheck,
     ZeroWidthSpaceCheck,
 )
 from weblate.formats.helpers import CONTROLCHARS_TRANS
 from weblate.trans.autofixes.base import AutoFix
+
+if TYPE_CHECKING:
+    from weblate.trans.models import Unit
 
 
 class ReplaceTrailingDotsWithEllipsis(AutoFix):
@@ -26,7 +33,9 @@ class ReplaceTrailingDotsWithEllipsis(AutoFix):
     def get_related_checks():
         return [EndEllipsisCheck()]
 
-    def fix_single_target(self, target, source, unit):
+    def fix_single_target(
+        self, target: str, source: str, unit: Unit
+    ) -> tuple[str, bool]:
         if source and source[-1] == "…" and target.endswith("..."):
             return f"{target[:-3]}…", True
         return target, False
@@ -42,7 +51,9 @@ class RemoveZeroSpace(AutoFix):
     def get_related_checks():
         return [ZeroWidthSpaceCheck()]
 
-    def fix_single_target(self, target, source, unit):
+    def fix_single_target(
+        self, target: str, source: str, unit: Unit
+    ) -> tuple[str, bool]:
         if unit.translation.language.base_code == "km":
             return target, False
         if "\u200b" not in source and "\u200b" in target:
@@ -56,7 +67,9 @@ class RemoveControlChars(AutoFix):
     fix_id = "control-chars"
     name = gettext_lazy("Control characters")
 
-    def fix_single_target(self, target, source, unit):
+    def fix_single_target(
+        self, target: str, source: str, unit: Unit
+    ) -> tuple[str, bool]:
         result = target.translate(CONTROLCHARS_TRANS)
         return result, result != target
 
@@ -67,9 +80,12 @@ class DevanagariDanda(AutoFix):
     fix_id = "devanadari-danda"
     name = gettext_lazy("Devanagari danda")
 
-    def fix_single_target(self, target, source, unit):
+    def fix_single_target(
+        self, target: str, source: str, unit: Unit
+    ) -> tuple[str, bool]:
         if (
             unit.translation.language.is_base(("hi", "bn", "or"))
+            and "_Latn" not in unit.translation.language.code
             and source.endswith(".")
             and target.endswith((".", "\u09f7", "|"))
         ):
@@ -87,14 +103,19 @@ class PunctuationSpacing(AutoFix):
     def get_related_checks():
         return [PunctuationSpacingCheck()]
 
-    def fix_single_target(self, target, source, unit):
+    def fix_single_target(
+        self, target: str, source: str, unit: Unit
+    ) -> tuple[str, bool]:
         if (
             unit.translation.language.is_base(("fr", "br"))
             and unit.translation.language.code != "fr_CA"
             and "ignore-punctuation-spacing" not in unit.all_flags
         ):
             # Fix existing
-            new_target = re.sub(FRENCH_PUNCTUATION_FIXUP_RE, "\u202f\\2", target)
+            new_target = re.sub(FRENCH_PUNCTUATION_FIXUP_RE_NBSP, "\u00a0\\2", target)
+            new_target = re.sub(
+                FRENCH_PUNCTUATION_FIXUP_RE_NNBSP, "\u202f\\2", new_target
+            )
             # Do not add missing as that is likely to trigger issues with other content
             # such as URLs or Markdown syntax.
             return new_target, new_target != target