+++ weblate/trans/views/hooks.py
@@ -425,8 +425,7 @@ def gitlab_hook_helper(data, request: AuthenticatedHttpRequest):
         data["repository"]["homepage"],
     ]
     full_name = ssh_url.split(":", 1)[1]
-    if full_name.endswith(".git"):
-        full_name = full_name[:-4]
+    full_name = full_name.removesuffix(".git")
 
     return {
         "service_long_name": "GitLab",